Updated and corrected multi-node hot water storage and lorentz-efficiency heat pump

+++ src/energy_system_control/core/registry.py
@@ -33,9 +33,9 @@ class SimulationData:
         self.controllers = None
 
     def create_empty_datasets(self, time_vector, signal_registry_ports, signal_registry_controllers, signal_registry_sensors):
-        self.ports = np.empty((len(time_vector), len(signal_registry_ports._col_to_key)), dtype=float)
-        self.controllers = np.empty((len(time_vector), len(signal_registry_controllers._col_to_key)), dtype=float)
-        self.sensors = np.empty((len(time_vector), len(signal_registry_sensors._col_to_key)), dtype=float)
+        self.ports = np.empty((len(time_vector), len(signal_registry_ports._col_to_key)), dtype=np.float32)
+        self.controllers = np.empty((len(time_vector), len(signal_registry_controllers._col_to_key)), dtype=np.float32)
+        self.sensors = np.empty((len(time_vector), len(signal_registry_sensors._col_to_key)), dtype=np.float32)
 
     def to_dataframe(self, time_vector, signal_registry_ports, signal_registry_controllers, signal_registry_sensors):
         columns = [f'{key.main_key}:{key.secondary_key}' for key in signal_registry_ports._col_to_key]
